backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleak>=2.1.1",
    "fastapi>=0.128.0",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
